Added "id" field to post and implemented related functionality
The following changes were made in this commit:
- An "id" field has been added to the database for each entry
- The logic of displaying and managing the "id" field was implemented in the console interface
- Updated documentation and comments for relevant functions

An "id" field has been added to uniquely identify each entry in the database. This allows you to more efficiently manage records, edit and delete them by unique identifier. All changes have been successfully tested and are ready to be included in the main branch of the project.